UI: Team view code separation, cleaning dummy data files
